Add files via upload
Unzip the Img.zip file first to take a look at the image. These images is from the chars74k dataset
The dataLoader.py file for creating the allData.csv, then will be processed in the Character_Recognition.py file 
all.txt file is a file that contains the names and path to access the images.